django/devbot/requirements/base.txt
gunicorn==19.1.1
distribute==0.7.3
Django==1.7.3
Pillow==2.6.1
boto==2.33.0
celery==3.1.8
dj-database-url==0.2.2
django-extensions==1.4.4
django-grappelli==2.6.1
django-require==1.0.3
django-storages==1.1.8
django-tastypie==0.12.1
mimeparse==0.1.3
psycopg2==2.5.1
pytz==2013.7